mixed: move scalar products to separate file
