+++ opm/simulators/linalg/mixed/SolverAdapter.hpp
@@ -9,215 +9,16 @@
 #include <dune/istl/operators.hh>
 #include <dune/istl/solvers.hh>
 #include <dune/istl/owneroverlapcopy.hh>
-
 #include <dune/istl/matrixindexset.hh>
 
 #include <opm/simulators/linalg/mixed/MatrixWrapper.hpp>
-
+#include <opm/simulators/linalg/ScalarProducts.hpp>
 
 
 namespace Dune
 {
-//#include <dune/istl/solvers.hh>
 
-
-//! @brief Optimized sequential scalar product.
-//!
-//! @tparam Vector block-vector class with data stored as contiguous double array
-template<class Vector>
-class SeqOptmizedProduct : public Dune::SeqScalarProduct<Vector>
-{
-public:
-
-    // extract block size
-    static constexpr auto block_size = Vector::block_type::dimension;
-
-    // Compute the dot product <vx, vy>
-    double dot(const Vector& vx, const Vector& vy) const override
-    {
-        // access underlying data
-        double const *x = &vx[0][0];
-        double const *y = &vy[0][0];
-
-        // total array length
-        int NN = block_size*vx.N();
-
-        // unroll loop in multiples of 8
-        int n=NN/8;
-        int N=8*n;
-        double agg[8];
-        for(int i=0;i<8;i++) agg[i]=0.0;
-        for(int i=0;i<N;i+=8) for(int j=0;j<8;j++) agg[j]+=x[i+j]*y[i+j];
-        for(int j=0;j<4;j++) agg[j]+=agg[j+4];
-        for(int j=0;j<2;j++) agg[j]+=agg[j+2];
-        for(int j=0;j<1;j++) agg[j]+=agg[j+1];
-
-        // trailing end
-        for(int j=N;j<NN;j++) agg[0]+=x[j]*y[j];
-
-        return agg[0];
-    }
-
-    // Compute the norm ||x||
-    double norm(const Vector& x) const override {
-        return std::sqrt(this->dot(x, x));
-    }
-};
-
-
-template<class Vector, class Comm>
-class GhostLastScalarProduct : public ScalarProduct<Vector>
-{
-    public:
-
-    static constexpr auto block_size = Vector::block_type::dimension;
-
-    //! \brief The type of the vector to compute the scalar product on.
-    //!
-    //! E.g. BlockVector or another type fulfilling the ISTL
-    //! vector interface.
-
-    /*!
-    * \param com The communication object for syncing overlap and copy
-    * data points.
-    * \param cat parallel solver category (nonoverlapping or overlapping)
-    */
-    GhostLastScalarProduct (std::shared_ptr<const Comm> com, SolverCategory::Category cat)
-        : _communication(com), _category(cat)
-    {
-        count_ = getLocalCount();
-        int verify = verifyLocalCount();
-        if (count_ != verify) OPM_THROW(std::runtime_error, "Inconsistent local node count!!\n");
-    }
-
-    /*!
-    * \param com The communication object for syncing overlap and copy
-    * data points.
-    * \param cat parallel solver category (nonoverlapping or overlapping)
-    * \note if you use this constructor you have to make sure com stays alive
-    */
-    GhostLastScalarProduct (const Comm& com, SolverCategory::Category cat)
-        : GhostLastScalarProduct(stackobject_to_shared_ptr(com), cat)
-    {}
-
-    /*! \brief Dot product of two vectors.
-    It is assumed that the vectors are consistent on the interior+border
-    partition.
-    */
-    void setLocalCount(int count){ count_ = count; }
-
-    virtual double dot (const Vector& vx, const Vector& vy) const override
-    {
-
-        // access underlying data
-        double const *x = &vx[0][0];
-        double const *y = &vy[0][0];
-
-        // total array length
-        int NN = block_size*count_;
-
-        //double result(0);
-
-        // unroll loop in multiples of 8
-        int n=NN/8;
-        int N=8*n;
-        double agg[8];
-        for(int i=0;i<8;i++) agg[i]=0.0;
-        for(int i=0;i<N;i+=8) for(int j=0;j<8;j++) agg[j]+=x[i+j]*y[i+j];
-        for(int j=0;j<4;j++) agg[j]+=agg[j+4];
-        for(int j=0;j<2;j++) agg[j]+=agg[j+2];
-        for(int j=0;j<1;j++) agg[j]+=agg[j+1];
-
-        // trailing end
-        for(int j=N;j<NN;j++) agg[0]+=x[j]*y[j];
-
-        // Global summation
-        auto cc = _communication->communicator();
-        double result = cc.sum(agg[0]);
-        return result;
-    }
-
-    /*! \brief Norm of a right-hand side vector.
-    The vector must be consistent on the interior+border partition
-    */
-    virtual double norm (const Vector& x) const override
-    {
-        return sqrt(dot(x,x));
-    }
-
-    //! Category of the scalar product (see SolverCategory::Category)
-    virtual SolverCategory::Category category() const override
-    {
-        return _category;
-    }
-
-    private:
-    std::shared_ptr<const Comm> _communication;
-    SolverCategory::Category _category;
-    int count_;
-
-    //int getLocalCount(const Comm& comm) const
-    int  getLocalCount() const
-    {
-        int count = 0;
-        // Loop over index set
-        auto indexSet = _communication->indexSet();
-        for (auto idx = indexSet.begin(); idx!=indexSet.end(); ++idx) {
-            if (idx->local().attribute()==1) count++; // count non-local indices
-        }
-        return count;
-    }
-
-    int verifyLocalCount() const
-    {
-        auto indexSet = _communication->indexSet();
-
-        size_t is = 0;
-        // Loop over index set
-        for (auto idx = indexSet.begin(); idx!=indexSet.end(); ++idx) {
-            //Only take "owner" indices
-            if (idx->local().attribute()==1) {
-                //get local index
-                auto loc = idx->local().local();
-                // if loc is higher than "old interior size", update it
-                if (loc > is) {
-                    is = loc;
-                }
-            }
-        }
-        return is + 1; //size is plus 1 since we start at 0
-    }
-
-};
-
-
-
-/*
-//! @brief Generalized mixed precision operator interface
-//!
-//! @tparam Matrix the block-matrix used by linear operator
-//! @tparam Vector the block-vector used by linear operator
-//! @tparam Comm the communicator used by linear operator
-template <class Matrix, class Vector, class Comm>
-struct MixedOperator
-{
-    using type = Dune::OverlappingSchwarzOperator<Matrix, Vector, Vector, Comm>;
-};
-
-//! @brief Generalized mixed precision operator interface
-//!
-//! @tparam Matrix the block-matrix used by linear operator
-//! @tparam Vector the block-vector used by linear operator
-template <class Matrix, class Vector>
-struct MixedOperator<Matrix, Vector, Dune::Amg::SequentialInformation>
-{
-    using type = Dune::MatrixAdapter<Matrix, Vector, Vector>;
-    //using type = Dune::AssembledLinearOperator<Matrix,Vector,Vector>;
-    //using type = Opm::WellModelMatrixAdapter<Matrix, Vector, Vector>;
-};
-*/
-
-//! @brief Wraps mixed precision
+//! @brief Adapts BiCGSTAB to mixed precision
 //!
 //! @tparam Comm the communicator passed to FlexibleLinearSolver
 //! @tparam Operator the linear operator passed to FlexibleLinearSolver
@@ -232,7 +33,6 @@ class MixedBiCGSTABSolver:public InverseOperator<Vector, Vector>
 
     static constexpr auto block_size = Vector::block_type::dimension;
     using MixedMatrixType      = Opm::MixedMatrixWrapper<Vector>;
-    //using MixedOperatorType    = MixedOperator<MixedMatrixType, Vector, Comm>::type;
 
     //! @brief constructor
     //!
